invoice-classifier-feature-aiohttp-web-app/requirements.txt
aiohttp==3.8.3
aiohttp-jinja2==1.5.1
aiosignal==1.3.1
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
bce-python-sdk==0.8.83
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
boto3==1.26.147
botocore==1.29.147
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.3
confection==0.0.4
contourpy==1.0.7
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.35
datefinder==0.7.3
decorator==5.1.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
flask-babel==3.1.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.0
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy-loader==0.2
lit==16.0.5.post0
lmdb==1.4.1
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
numpy==1.24.1
opencv-contrib-python==4.5.4.60
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==2.0.2
pathy==0.10.1
pdf2docx==0.5.6
Pillow==9.4.0
premailer==3.10.0
preshed==3.0.8
protobuf==3.20.0
psutil==5.9.5
pyclipper==1.3.0.post4
pycryptodome==3.18.0
pydantic==1.10.8
PyMuPDF==1.20.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytorch-transformers==1.2.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.1.0
rarfile==4.0
regex==2022.10.31
requests==2.31.0
s3transfer==0.6.1
sacremoses==0.0.53
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.12
termcolor==2.3.0
thinc==8.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.3
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.28.1
triton==2.0.0
typer==0.7.0
typing-extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.114
visualdl==2.5.3
wasabi==1.1.1
Werkzeug==2.3.4
xgboost==1.7.5
yarl==1.9.2
zipp==3.15.0
